Improving error handling of ditool and adding support for DiskCopyII images.
git-svn-id: https://svn.code.sf.net/p/previous/code/branches/branch_filesharing@1704 9ba66e4b-2233-44b8-a46a-7e93b2105c9e

+++ src/ditool/ufs.c
@@ -39,7 +39,7 @@ struct ufs_t* ufs_init(struct part_t* part) {
                     ufs->cacheBlockNo[i] = BLOCK_INVALID;
                 }
             } else {
-                printf("No valid file system found (%X).\n", ntohl(superblock->fs_magic));
+                printf("No valid file system found (%X).\n", (uint32_t)ntohl(superblock->fs_magic));
             }
         } else {
             printf("Can't read super block.\n");
